backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
bcrypt
PyJWT
python-dotenv
cryptography